refactoring

+++ WeatherGrabber/Models/ForecastGrabberModel.cs
@@ -3,7 +3,7 @@
     public class ForecastGrabberModel : IEquatable<ForecastGrabberModel>
     {
         public DateTime Date { get; set; }
-        public double Temperature { get; set; }
+        public int Temperature { get; set; }
         public string Condition { get; set; } = null!;
 
         public bool Equals(ForecastGrabberModel? other)
